wip: extract shared types
